# incubator-backend/openapi.yaml
openapi: 3.0.3
info:
  title: IT Youth Talent Incubator API
  description: |
    # IT Youth Talent Incubator Backend API
    
    Welcome to the comprehensive API documentation for the IT Youth Talent Incubator platform.
    This platform connects talented IT students in Ghana with companies seeking skilled professionals.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **User Management**: Support for Students, Companies, and Admins with distinct profiles
    - **Job Posting & Management**: Companies can post, manage, and track jobs
    - **Application System**: Students can apply to jobs and track application status
    - **Admin Panel**: Comprehensive admin tools for user approval and platform management
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 3-5 requests per 15 minutes
    
    ## Pagination
    List endpoints support pagination with the following query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 10, max: 100)
    
  version: 1.0.0
  contact:
    name: IT Youth Talent Incubator
    email: support@itfy-incubator.com
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation Error"
            error: "Some required fields are missing. Please fill them all."
...
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items per page
...
paths:
...
  /users/companies/me/jobs/stats:
    get:
      tags:
        - Companies
        - Jobs
      summary: Get job posting statistics
      description: Retrieve statistics for company's job postings
      operationId: getCompanyJobStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Job statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_jobs:
                        type: integer
                        example: 25
                      active_jobs:
                        type: integer
                        example: 10
                      draft_jobs:
                        type: integer
                        example: 3
                      closed_jobs:
                        type: integer
                        example: 12
                      total_applications:
                        type: integer
                        example: 150
                      pending_applications:
                        type: integer
                        example: 45
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: List all active jobs
      description: Get a paginated list of all published and active jobs
      operationId: listActiveJobs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: job_type
          schema:
            type: string
            enum: [full-time, part-time, internship, contract, remote, hybrid]
        - in: query
          name: experience_level
          schema:
            type: string
            enum: [entry, junior, mid, senior, lead, executive]
        - in: query
          name: location
          schema:
            type: string
      responses:
        '200':
          description: Jobs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Jobs
      summary: Create a new job (Company only)
      description: Create a new job posting
      operationId: createJob
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/featured:
    get:
      tags:
        - Jobs
      summary: Get featured jobs
      description: Retrieve list of featured job postings
      operationId: getFeaturedJobs
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Featured jobs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/recent:
    get:
      tags:
        - Jobs
      summary: Get recent jobs
      description: Retrieve recently posted jobs
      operationId: getRecentJobs
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Recent jobs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/trending:
    get:
      tags:
        - Jobs
      summary: Get trending jobs
      description: Retrieve jobs with high engagement
      operationId: getTrendingJobs
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Trending jobs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/search:
    get:
      tags:
        - Jobs
      summary: Search jobs
      description: Search jobs by title, description, or skills
      operationId: searchJobs
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/filter:
    get:
      tags:
        - Jobs
      summary: Filter jobs
      description: Filter jobs by multiple criteria
      operationId: filterJobs
      parameters:
        - in: query
          name: job_type
          schema:
            type: array
            items:
              type: string
        - in: query
          name: experience_level
          schema:
            type: array
            items:
              type: string
        - in: query
          name: skills
          schema:
            type: array
            items:
              type: string
        - in: query
          name: location
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Filtered jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/my-jobs:
    get:
      tags:
        - Jobs
      summary: Get my jobs (Company only)
      description: Retrieve all jobs posted by the authenticated company
      operationId: getMyJobs
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, published, paused, closed]
      responses:
        '200':
          description: Jobs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/my-jobs/stats:
    get:
      tags:
        - Jobs
      summary: Get my job statistics (Company only)
      description: Get statistics for company's job postings
      operationId: getMyJobStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalJobs:
                        type: integer
                      publishedJobs:
                        type: integer
                      draftJobs:
                        type: integer
                      pausedJobs:
                        type: integer
                      closedJobs:
                        type: integer
                      totalViews:
                        type: integer
                      totalApplications:
                        type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/{id}:
    get:
      tags:
        - Jobs
      summary: Get job by ID
      description: Retrieve detailed information about a specific job
      operationId: getJobById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Job ID
      responses:
        '200':
          description: Job retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Jobs
      summary: Update job (Company only)
      description: Update a job posting
      operationId: updateJob
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
      responses:
        '200':
          description: Job updated successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Jobs
      summary: Delete job (Company only)
      description: Soft delete a job posting
      operationId: deleteJob
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /jobs/slug/{slug}:
    get:
      tags:
        - Jobs
      summary: Get job by slug
      description: Retrieve job using URL-friendly slug
      operationId: getJobBySlug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          description: Job slug
      responses:
        '200':
          description: Job retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /jobs/{id}/publish:
    post:
      tags:
        - Jobs
      summary: Publish job (Company only)
      description: Publish a draft job posting
      operationId: publishJob
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job published successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /jobs/{id}/close:
    post:
      tags:
        - Jobs
      summary: Close job (Company only)
      description: Close a published job posting
      operationId: closeJob
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job closed successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/{id}/pause:
    post:
      tags:
        - Jobs
      summary: Pause job (Company only)
      description: Temporarily pause a job posting
      operationId: pauseJob
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job paused successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/{id}/analytics:
    get:
      tags:
        - Jobs
      summary: Get job analytics (Company only)
      description: Retrieve analytics data for a specific job
      operationId: getJobAnalytics
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      views:
                        type: integer
                      applications:
                        type: integer
                      viewsOverTime:
                        type: array
                        items:
                          type: object
                      applicationsOverTime:
                        type: array
                        items:
                          type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /jobs/bulk-update:
    post:
      tags:
        - Jobs
      summary: Bulk update job status (Company only)
      description: Update status for multiple jobs at once
      operationId: bulkUpdateStatus
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobIds:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [draft, published, paused, closed]
      responses:
        '200':
          description: Jobs updated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...